Add concurrency property in `Deploy master` action
We use a third-party action to upload our build folder to our hosting. It also uploads a file keeping track of modified files in order to improve upload time and bandwidth usage. If that file was modified by multiple parties at the same time, it could result in the action not working anymore, needing manual intervention to fix the issue.
With the `concurrency` property, we can disallow multiple instances of this action to be running at the same time, and new instances will be waiting for the previous ones to finish.