Fix pages always being redirected to index.html
This was happening because of the order of rewrite rules in the `.htaccess` file. Requests were correctly redirected to the `index.html` file so that react-router could handle routing, but after updating to react-router 6, redirecting to https/www after rewriting to `index.html` would lose information about what page the request was trying to access.
Instead, rewrite to https/www first and then redirect to `index.html`.